buildvnc -v: Fail if subscript fails
If -v was passed to buildvnc (which our CI system always does), then the
exit status of buildvnc was the exit status of tee, not the exit status
of buildvnc.$PLATFORM.  Effectively, that meant that errors in the
official build jobs were not communicated to GitHub Actions.
'set -o pipefail' causes the exit status of buildvnc to be the logical
AND of the exit statuses of buildvnc.$PLATFORM and tee.